+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
+- グローバル設定に注文変更時通知設定を追加
+
 ### Changed
 
 ### Deprecated